feat: Add bootstrap-k8s manifests and sealed-secrets
Add bootstrap-k8s manifests and sealed-secrets to the repository.
This includes the creation of the 'manifests.yaml' file and the 'sealed-secrets.yaml' file.
The 'manifests.yaml' file contains the configuration for the bootstrap application set,
while the 'sealed-secrets.yaml' file is used for managing sealed secrets in the argocd namespace.
Also, a new secret named 'mysecret' has been added to the 'mysecret.yaml' file in the bootstrap manifests.

Refactor the sync-wave annotation in the 'sealed-secrets.yaml' file from "40" to "10" to align with the desired synchronization order.

Refactor the targetRevision in the bootstrap-k8s/values.yaml file to 'main' and add the crds.enabled field in the cert-manager/values.yaml file.

Refactor the cert-manager.yaml and values.yaml files to update the configuration of cert-manager.

feat. cert-manager